requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
phonenumbers==8.13.43
pytz==2024.1
